Unify the filters definitions
