+++ src/Sylius/Bundle/ApiBundle/Doctrine/ORM/Filter/ProductVariantCatalogPromotionFilter.php
@@ -11,27 +11,30 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\Filter\Doctrine;
+namespace Sylius\Bundle\ApiBundle\Doctrine\ORM\Filter;
 
 use ApiPlatform\Api\IriConverterInterface;
 use ApiPlatform\Doctrine\Orm\Filter\AbstractFilter;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
 use ApiPlatform\Metadata\Operation;
+use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
 use Psr\Log\LoggerInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 
-final class ChannelsAwareChannelFilter extends AbstractFilter
+final class ProductVariantCatalogPromotionFilter extends AbstractFilter
 {
     public function __construct(
-        private readonly IriConverterInterface $iriConverter,
+        private IriConverterInterface $iriConverter,
         ManagerRegistry $managerRegistry,
+        ?RequestStack $requestStack = null,
         ?LoggerInterface $logger = null,
         ?array $properties = null,
         ?NameConverterInterface $nameConverter = null,
     ) {
-        parent::__construct($managerRegistry, $logger, $properties, $nameConverter);
+        parent::__construct($managerRegistry, $requestStack, $logger, $properties, $nameConverter);
     }
 
     protected function filterProperty(
@@ -43,26 +46,37 @@ final class ChannelsAwareChannelFilter extends AbstractFilter
         ?Operation $operation = null,
         array $context = [],
     ): void {
-        if ('channel' !== $property) {
+        if ('catalogPromotion' !== $property) {
             return;
         }
 
-        $channel = $this->iriConverter->getResourceFromIri($value);
+        $catalogPromotion = $this->iriConverter->getResourceFromIri($value);
+
         $parameterName = $queryNameGenerator->generateParameterName($property);
+        $channelPricingJoinAlias = $queryNameGenerator->generateJoinAlias('channelPricing');
+        $appliedPromotionJoinAlias = $queryNameGenerator->generateJoinAlias('appliedPromotion');
         $rootAlias = $queryBuilder->getRootAliases()[0];
+
         $queryBuilder
-            ->andWhere(sprintf(':%s MEMBER OF %s.channels', $parameterName, $rootAlias))
-            ->setParameter($parameterName, $channel)
+            ->leftJoin(sprintf('%s.channelPricings', $rootAlias), $channelPricingJoinAlias)
+            ->innerJoin(
+                sprintf('%s.appliedPromotions', $channelPricingJoinAlias),
+                $appliedPromotionJoinAlias,
+                Join::WITH,
+                sprintf('%s = :%s', $appliedPromotionJoinAlias, $parameterName),
+            )
+            ->setParameter($parameterName, $catalogPromotion)
         ;
     }
 
     public function getDescription(string $resourceClass): array
     {
         return [
-            'channel' => [
+            'catalogPromotion' => [
                 'type' => 'string',
                 'required' => false,
-                'property' => 'channels',
+                'property' => null,
+                'description' => 'Get a collection of product variants with applied catalog promotion',
                 'schema' => [
                     'type' => 'string',
                 ],
