+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/admin/CatalogPromotion.xml
@@ -36,7 +36,7 @@
                     <filter>sylius_api.search_filter.admin.catalog_promotion</filter>
                     <filter>sylius_api.boolean_filter.admin.catalog_promotion</filter>
                     <filter>sylius_api.date_filter.admin.catalog_promotion</filter>
-                    <filter>sylius_api.search_filter.channel_aware</filter>
+                    <filter>sylius_api.channel_search_filter.admin.channel_aware</filter>
                     <filter>sylius_api.order_filter.admin.catalog_promotion</filter>
                 </filters>
             </operation>
