+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/admin/PaymentMethod.xml
@@ -34,7 +34,7 @@
                 </normalizationContext>
                 <filters>
                     <filter>sylius_api.order_filter.admin.payment_method</filter>
-                    <filter>sylius_api.order_filter.translation_order_name_and_locale</filter>
+                    <filter>sylius_api.name_search_order_by_locale_filter.admin.translatable</filter>
                     <filter>sylius_api.search_filter.admin.payment_method</filter>
                     <filter>sylius_api.boolean_filter.admin.payment_method</filter>
                 </filters>
