+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/admin/Product.xml
@@ -30,8 +30,8 @@
                 </normalizationContext>
                 <filters>
                     <filter>sylius_api.search_filter.admin.product</filter>
-                    <filter>sylius_api.search_filter.channel_aware</filter>
-                    <filter>sylius_api.order_filter.translation_order_name_and_locale</filter>
+                    <filter>sylius_api.search_filter.admin.channel_aware</filter>
+                    <filter>sylius_api.name_search_order_by_locale_filter.admin.translatable</filter>
                     <filter>sylius_api.order_filter.admin.product</filter>
                 </filters>
                 <order>
