+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/admin/ShippingMethod.xml
@@ -35,7 +35,7 @@
                 <filters>
                     <filter>sylius_api.exists_filter.admin.shipping_method</filter>
                     <filter>sylius_api.order_filter.admin.shipping_method</filter>
-                    <filter>sylius_api.order_filter.translation_order_name_and_locale</filter>
+                    <filter>sylius_api.name_search_order_by_locale_filter.admin.translatable</filter>
                 </filters>
                 <order>
                     <values>
