+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/shop/Product.xml
@@ -31,9 +31,9 @@
                 <filters>
                     <filter>sylius_api.search_filter.shop.product</filter>
                     <filter>sylius_api.order_filter.shop.product</filter>
-                    <filter>sylius_api.filter.product_price_order</filter>
-                    <filter>sylius_api.order_filter.translation_order_name_and_locale</filter>
-                    <filter>sylius_api.filter.taxon</filter>
+                    <filter>sylius_api.price_filter.shop.product</filter>
+                    <filter>sylius_api.name_search_order_by_locale_filter.shop.translatable</filter>
+                    <filter>sylius_api.taxon_filter.shop.product</filter>
                 </filters>
                 <order>
                     <values>
